Renamed elevationbehavior.py to elevation.py
