+++ kivymd/theme_picker.py
@@ -22,7 +22,7 @@ from kivy.utils import get_color_from_hex
 from kivymd.backgroundcolorbehavior import SpecificBackgroundColorBehavior
 from kivymd.button import MDIconButton
 from kivymd.color_definitions import colors
-from kivymd.elevationbehavior import RectangularElevationBehavior
+from kivymd.elevation import RectangularElevationBehavior
 from kivymd.theming import ThemableBehavior
 
 Builder.load_string('''
